backend/vss/requirements.txt
eve
eve-sqlalchemy
Flask-SQLAlchemy
#py-bcrypt
Eve-docs
# Mocker
#eve-mocker
#sure
pyresttest
future
requests
